requirements.txt
fastapi==0.104.1
python-telegram-bot==20.7
uvicorn==0.24.0
python-dotenv==1.0.0
